Add javadoc, rename DruidErrorCode, and address other minor comments
